Added unprojection so we can get the mouse coords correctly on a gui. Still needs to be optimizied. Added a notice to the main fail, about to add this to all files.

+++ src/display.cpp
@@ -154,7 +154,7 @@ void Display::OnMouseMotion(){
 	if( Mouse_GetButtonState(0) ){ //dragging
 		gui->Move( Mouse_GetChangeX(), Mouse_GetChangeY() );
 	} else {
-		gui->HitTest( Mouse_GetX(), Mouse_GetY() );
+		gui->HitTest( (float)Mouse_GetX(), (float)Mouse_GetY() );
 	}
 }
